+++ gallery-backend/src/public/db/tree_snapshot/read_scrollbar.rs
@@ -2,14 +2,15 @@ use std::time::Instant;
 
 use super::TreeSnapshot;
 use crate::{
-    public::db::tree_snapshot::read_tree_snapshot::MyCow, public::structure::response::row::ScrollBarData,
+    public::db::tree_snapshot::read_tree_snapshot::MyCow,
+    public::structure::response::row::ScrollBarData,
 };
 
 use chrono::{Datelike, TimeZone, Utc};
 use redb::ReadableTable;
 
 impl TreeSnapshot {
-    pub fn read_scrollbar(&'static self, timestamp: u128) -> Vec<ScrollBarData> {
+    pub fn read_scrollbar(&'static self, timestamp: i64) -> Vec<ScrollBarData> {
         let start_time = Instant::now();
         let tree_snapshot = self.read_tree_snapshot(&timestamp).unwrap();
         let mut scroll_bar_data_vec = Vec::new();
